drm/msm/sde: Skip heavy autorefresh checks when it's not enabled
These heavy checks for seeing if autorefresh is enabled are unneeded
when the autorefresh config is disabled. These checks are performed on
every display commit and show up as using a significant amount of CPU
time in perf top. Skip them when it's unnecessary in order to improve
display rendering performance.

Change-Id: Ibc806c6cc0f7413d2862644937438720a281ad66
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: HeroBuxx <me@herobuxx.me>